client/requirement.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pillow==11.2.1
qrcode==8.2
requests==2.32.4
urllib3==2.5.0
